+++ library/src/main/java/me/denley/courier/compiler/Recipient.java
@@ -5,13 +5,23 @@ import javax.lang.model.element.ElementKind;
 public class Recipient {
 
     public final String recipientName;
-    public final ElementKind deliveryType;
     public final String payloadType;
+
+    public final ElementKind deliveryType;
+    public boolean hasNodeParameter = false;
+
 
     public Recipient(String name, ElementKind type, String payload) {
         this.recipientName = name;
         this.deliveryType = type;
         this.payloadType = payload;
+    }
+
+    public Recipient(String name, ElementKind type, String payload, boolean hasNodeParameter) {
+        this.recipientName = name;
+        this.deliveryType = type;
+        this.payloadType = payload;
+        this.hasNodeParameter = hasNodeParameter;
     }
 
     public void writeDataBindingTo(StringBuilder builder) {
@@ -31,7 +41,11 @@ public class Recipient {
         builder.append(recipientName);
 
         if(deliveryType==ElementKind.METHOD){
-            builder.append("(").append(sourceName).append(")");
+            builder.append("(").append(sourceName);
+            if(hasNodeParameter) {
+                builder.append(", node");
+            }
+            builder.append(")");
         } else {
             builder.append(" = ").append(sourceName);
         }